requirements.txt
requests
python-dotenv
pytest
sphinx
sphinxcontrib-napoleon